+++ skibaProj/src/main/java/parentPage/ParentPage.java
@@ -1,21 +1,14 @@
 package parentPage;
 
-import libs.ActionsWithOurElements;
-import org.apache.log4j.Logger;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.support.PageFactory;
+
 
 public class ParentPage {
     protected WebDriver webDriver;
-    protected Logger logger = Logger.getLogger(getClass());
-    protected ActionsWithOurElements actionsWithOurElements;
 
     public ParentPage(WebDriver webDriver) {
         this.webDriver = webDriver;
-        PageFactory.initElements(webDriver, this);
-        actionsWithOurElements = new ActionsWithOurElements(webDriver);
     }
-
 
 
 }
