+++ skibaProj/src/test/java/abstractParentTest/AbstractParentTest.java
@@ -1,13 +1,10 @@
 package abstractParentTest;
 
 import org.junit.After;
-import org.junit.Assert;
 import org.junit.Before;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
-import pages.HomePage;
 import pages.LoginPage;
-import pages.SparePage;
 
 import java.io.File;
 import java.util.concurrent.TimeUnit;
@@ -15,10 +12,8 @@ import java.util.concurrent.TimeUnit;
 public class AbstractParentTest {
     WebDriver webDriver;
     protected LoginPage loginPage;
-    protected HomePage homePage;
-    protected SparePage sparePage;
 
-    @Before
+    @Before //this method will setUp our tests before every test
     public void setUp(){
         File file = new File("./src/drivers/chromedriver.exe");
         System.setProperty("webdriver.chrome.driver", file.getAbsolutePath());
@@ -28,23 +23,10 @@ public class AbstractParentTest {
         webDriver.manage().timeouts().implicitlyWait(5, TimeUnit.SECONDS);
 
         loginPage = new LoginPage(webDriver);
-        homePage = new HomePage(webDriver);
-        sparePage= new SparePage(webDriver);
-
-
     }
 
     @After
     public void tearDown(){
         webDriver.quit();
-    }
-
-    protected void checkExpectedResult(String message, boolean actualResult) {
-        Assert.assertEquals(message, true, actualResult);
-    }
-
-    protected void checkExpectedResult(String message,boolean expectedResult
-            , boolean actualResult) {
-        Assert.assertEquals(message, expectedResult, actualResult);
     }
 }
