--- yerastovaProj/src/main/java/DTO/CurrencyValues.java
@@ -1,9 +1,0 @@
-package DTO;
-
-public class CurrencyValues {
-    public String ccy;
-    public String base_ccy;
-    public String buy;
-    public String sale;
-
-}
