--- yerastovaProj/src/main/java/libs/ConfigProperties.properties
@@ -1,22 +1,0 @@
-base_url=http://v3.test.itpmgroup.com
-TIME_FOR_DFFAULT_WAIT=10
-TIME_FOR_EXPLICIT_WAIT_LOW=15
-TIME_FOR_EXPLICIT_WAIT_HIGHT=30
-
-### Data file location
-# single test running : data/testData.xls
-# test suite running  : data/testSuiteData.xls
-DATA_FILE=.//src//main//java//data//testData.xls
-DATA_FILE_PATH=.//src//main//java//data//
-
-### DB drivers
-# DB2=com.ibm.db2.jcc.DB2Jcc
-Oracle=oracle.jdbc.driver.OracleDriver
-MySQL=com.mysql.jdbc.Driver
-sqlServer=com.microsoft.sqlserver.jdbc.SQLServerDriver
-
-
-## For QALight BD
-MySQL_DB=jdbc:mysql://185.25.118.54:3306/old_qa
-MySQL_DB_USER=old_qa
-MySQL_DB_PASSWORD=PB3Q2F6tdSD9cL3z
