--- yerastovaProj/src/main/java/pages/pageElements/WebDriverAware.java
@@ -1,8 +1,0 @@
-package pages.pageElements;
-
-import org.openqa.selenium.WebDriver;
-
-public interface WebDriverAware {
-    void setWebDriver(WebDriver driver);
-
-}
