+++ yerastovaProj/src/test/java/loginTest/LoginWithPageObject.java
@@ -1,0 +1,13 @@
+package loginTest;
+
+import abstractParentTest.AbstractParentTest;
+import org.junit.Test;
+
+public class LoginWithPageObject extends AbstractParentTest {
+
+    @Test
+    public void validLogin(){
+        loginPage.openPage();
+
+    }
+}
