--- yerastovaProj/src/test/java/loginTest/LoginWithPageObjectTest.java
@@ -1,36 +1,0 @@
-package loginTest;
-
-import abstractParentTest.AbstractParentTest;
-import io.qameta.allure.*;
-import org.junit.Test;
-@Epic("Allure examples")
-@Feature("Junit 4 support")
-public class LoginWithPageObjectTest extends AbstractParentTest {
-
-    @Description("Some detailed test description")
-    @Story("Base support for bdd annotations")
-    @Link("https://example.org")
-    @Link(name = "allure", type = "mylink")
-    @Issue("123")
-    @Issue("432")
-    @Severity(SeverityLevel.CRITICAL)
-    @Test
-    public void validLogin () {
-        loginPage.openPage();
-        loginPage.enterLoginIntoInputLogin("Student");
-        loginPage.enterPassIntoInputPassword("909090");
-        loginPage.clickOnButtonVhod();
-        checkExpectedResult("Avatar is not present",true, homePage.isAvatarDisplayed());
-
-    }
-    
-    @Test
-    public void invalidPassword(){
-        loginPage.openPage();
-        loginPage.enterLoginIntoInputLogin("Student");
-        loginPage.enterPassIntoInputPassword("qwerty");
-        loginPage.clickOnButtonVhod();
-        checkExpectedResult("Button Vhod is not present", true, loginPage.isSubmitButtonPresent());
-    }
-    
-}
