--- yerastovaProj/src/test/java/providersTest/ProvidersTest.java
@@ -1,20 +1,0 @@
-package providersTest;
-
-import abstractParentTest.AbstractParentTest;
-import org.junit.Test;
-
-public class ProvidersTest extends AbstractParentTest {
-    @Test
-    public void changeCheckBoxState(){
-        loginPage.fillingLoginFormAndSubmitIt("Student", "909090");
-
-        homePage.checkIsAvatarPresent();
-       // homePage.clickOnMenuDictionary();
-        homePage.clickOnSubMenuProviders();
-
-        providersPage.clickOnAddButton();
-
-        editProvidersPage.setStateToCheckboxPrivatePerson ("Check");
-        editProvidersPage.setStateToCheckboxIsOurFirm ("uncheck");
-    }
-}
