--- yerastovaProj/src/test/java/sparesTest/SparesTest.java
@@ -1,40 +1,0 @@
-package sparesTest;
-
-import abstractParentTest.AbstractParentTest;
-import libs.Utils;
-import org.junit.After;
-import org.junit.Test;
-
-public class SparesTest extends AbstractParentTest {
-    private final String spareName = "YerastovaSpare" + Utils.getDateAndTimeFormated();
-    @Test
-    public  void addNewSpare () {
-        loginPage.fillingLoginFormAndSubmitIt("Student", "909090");
-
-        homePage.checkCurrentUrl();
-        homePage.checkIsAvatarPresent();
-//        homePage.clickOnMenuDictionary();
-//        homePage.clickOnSubMenuSpares();
-        homePage.leftMenu.clickOnMenuDictionary();
-        homePage.leftMenu.clickOnSubMenuSpare();
-
-        sparePage.checkCurrentUrl();
-        sparePage.deleteSpareUntilPresent(spareName);
-
-        sparePage.clickOnAddButton();
-
-        editSparePage.checkCurrentUrl();
-        editSparePage.enterSpareNameIntoInput(spareName);
-        editSparePage.selectSpareTypeFromDropDown("Датчики");
-        editSparePage.clickOnSubmitButton();
-
-        sparePage.checkCurrentUrl();
-        checkExpectedResult("Can not find spare in list", true, sparePage.isSpareInList(spareName));
-
-    }
-
-    @After
-    public void deleteSpare () {
-        sparePage.deleteSpareUntilPresent(spareName);
-    }
-}
